+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/SyncedRealmTests.java
@@ -63,7 +63,7 @@ public class SyncedRealmTests extends BaseIntegrationTest {
         final SyncUser user = loginUser();
 
         SyncConfiguration config = new SyncConfiguration.Builder(user, Constants.USER_REALM)
-                .waitForInitialRemoteData()
+                .waitForInitialRemoteData(false)
                 .build();
 
         Realm realm = null;
@@ -102,7 +102,7 @@ public class SyncedRealmTests extends BaseIntegrationTest {
         SyncCredentials credentials = SyncCredentials.usernamePassword(UUID.randomUUID().toString(), "password", true);
         SyncUser user = SyncUser.login(credentials, Constants.AUTH_URL);
         SyncConfiguration config = new SyncConfiguration.Builder(user, Constants.USER_REALM)
-                .waitForInitialRemoteData()
+                .waitForInitialRemoteData(false)
                 .build();
 
         Realm realm = null;
@@ -124,7 +124,7 @@ public class SyncedRealmTests extends BaseIntegrationTest {
         SyncCredentials credentials = SyncCredentials.usernamePassword(UUID.randomUUID().toString(), "password", true);
         SyncUser user = SyncUser.login(credentials, Constants.AUTH_URL);
         final SyncConfiguration config = new SyncConfiguration.Builder(user, Constants.USER_REALM)
-                .waitForInitialRemoteData()
+                .waitForInitialRemoteData(false)
                 .build();
 
         for (int i = 0; i < 10; i++) {
@@ -160,7 +160,7 @@ public class SyncedRealmTests extends BaseIntegrationTest {
         SyncCredentials credentials = SyncCredentials.usernamePassword(UUID.randomUUID().toString(), "password", true);
         SyncUser user = SyncUser.login(credentials, Constants.AUTH_URL);
         final SyncConfiguration config = new SyncConfiguration.Builder(user, Constants.USER_REALM)
-                .waitForInitialRemoteData()
+                .waitForInitialRemoteData(false)
                 .build();
         Random randomizer = new Random();